requirements.txt
PyYAML==6.0.2
pandas==2.2.2
Jinja2==3.1.4
python-docx==1.1.2
scholarly==1.7.11
